requirements.txt
langchain
langchain-community
python-dotenv
langchain-google-genai
requests
beautifulsoup4
pandas
psutil
